(icds split) Remove a bunch of domains from ICDS report configs
Where the corresponding data sources aren't active for those domains.  It would
be just as easy to add the data sources to these domains, but I figure this is a
better default, and since these can't be working as-is anyways, it's a safe
enough option.  Also, these are all test environments.

Example report missing data source:
https://india.commcarehq.org/a/icds-sql/reports/configurable/static-icds-sql-static-mpr_5_child_health_cases/