(icds split) port stunting report to mobile dashboard
this commit does a number of different things, but I thought it would be
useful to have a single commit that contains the entire change of porting
one report. the individual parts are:

1. Moving the directive import into the shared file so it gets included
   in the mobile dashboard template
2. Using the templateProviderService to properly swap in mobile templates
3. Passing isMobile up and into the baseController
4. Adding sectionSlug for back button navigation
5. Adding serviceDataFunction and removing loadData to use the base class's
   implementation (which adds mobile support for loading both steps data)
6. Swapping templatePopup with getPopupData and leveraging the base class's
   implementation of that function, so it can also be used by
   displayMobilePopup

Though each of these changes could theoretically be their own standalone
commit, I'm going to do this for each report in a single commit to avoid
blowing up the git history.